feat: widen block numbers to u64 and reject malformed slices
Data::block_no combined a u32 slice number with a u32 blocks_per_slice
into a u32, so the coordinate pair could name far more blocks than the
result could hold: a forged slice_no overflowed the multiply, panicking
in debug and wrapping in release, which wrote the block at a garbage
offset. The value is now u64, which the field widths cannot overflow.

block_no also validates its inputs, since a zero blocks_per_slice
divided by zero and a block_in_slice beyond the slice aliased another
block's slot. Offsets are computed with checked_mul, which a forged
Hello carrying blocks_per_slice=u32::MAX can still overflow.

Receivers now track arrivals in a BlockCounter bitset and report
progress as Stats, which the rate controller will consume.

+++ crates/scuttlecast/src/receiver/reorderer.rs
@@ -3,8 +3,8 @@ use tokio::sync::mpsc;
 
 #[derive(Debug)]
 pub struct Reorderer {
-    next_block: u32,
-    pending: BTreeMap<u32, Vec<u8>>,
+    next_block: u64,
+    pending: BTreeMap<u64, Vec<u8>>,
     tx: mpsc::Sender<Vec<u8>>,
 }
 
@@ -21,7 +21,7 @@ impl Reorderer {
             rx,
         )
     }
-    pub async fn on_block(&mut self, block_no: u32, payload: Vec<u8>) {
+    pub async fn on_block(&mut self, block_no: u64, payload: Vec<u8>) {
         if block_no < self.next_block {
             return;
         }
